server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.9.2
openpyxl==3.1.5
